organizing effects and model
