+++ examples/weather/shared/src/model/active/favorites/mod.rs
@@ -1,2 +1,3 @@
+pub mod client;
 pub mod events;
 pub mod model;
