+++ examples/weather/shared/src/model/active/favorites/model.rs
