+++ examples/weather/shared/src/model/active/location.rs
