+++ examples/weather/shared/src/model/active/weather/client.rs
@@ -5,7 +5,7 @@ use crux_http::protocol::HttpRequest;
 use serde::{Deserialize, Serialize};
 
 use crate::effects::location::Location;
-use crate::model::weather::model::current_response::{CurrentWeatherResponse, WEATHER_URL};
+use crate::model::active::weather::model::current_response::{CurrentWeatherResponse, WEATHER_URL};
 
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
 pub enum WeatherError {
@@ -28,7 +28,7 @@ impl WeatherApi {
     /// Build an `HttpRequest` for testing purposes
     #[cfg(test)]
     pub fn build(location: Location, api_key: &str) -> HttpRequest {
-        use crate::model::weather::model::current_response::WEATHER_URL;
+        use crate::model::active::weather::model::current_response::WEATHER_URL;
 
         HttpRequest::get(WEATHER_URL)
             .query(&CurrentWeatherQuery {
