+++ examples/weather/shared/src/model/active/weather/model/current_response.rs
