+++ examples/weather/shared/src/model/active/weather/model/mod.rs
