+++ examples/weather/shared/src/model/active/weather/model/response_elements/wind.rs
